requirements.txt
yfinance==0.24.5
pandas==2.2.2
python-telegram-bot==20.10
